truncated models resnet 50 cifar lvl 9 All levels vgg19 and tiny_imagenet, All levels resnet50 and tiny_imagenet.

+++ train_CIFAR10.py
@@ -259,7 +259,7 @@ def train(epoch):
         if batch_idx == 0:
             t0 = time.time()
         outputs = net(inputs)
-        print("Labels: {}".format(targets.to(torch.LongTensor)))
+        print("Labels: {}".format(targets))
         print("\n")
         print("Labels shape {}".format(targets.shape))
         print("\n")
